cloudstore breakdown
